+++ spec/integration/greater_than_equal_spec.rb
@@ -15,12 +15,4 @@ describe "Definition.GreaterThanEqual" do
 
     it_behaves_like "it does not conform"
   end
-
-  context "with deprecated interface that had a typo" do
-    subject(:definition) do
-      Definition.GreaterThenEqual(5)
-    end
-
-    it_behaves_like "it conforms", 5
-  end
 end